Filter and sorting classes moved to new packages
