+++ src/test/java/com/es/phoneshop/model/comparartor/SortingComparatorTest.java
@@ -9,10 +9,10 @@ import java.util.Currency;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.es.phoneshop.comparator.SortingComparator;
+import com.es.phoneshop.comparator.filter.Filter;
 import com.es.phoneshop.enums.SortField;
 import com.es.phoneshop.enums.SortOrder;
-import com.es.phoneshop.model.comparator.SortingComparator;
-import com.es.phoneshop.model.filter.Filter;
 import com.es.phoneshop.model.product.Product;
 
 public class SortingComparatorTest {
